Refactor common components
