+++ client/src/components/Link/index.js
@@ -1,0 +1,3 @@
+import LoginLink from "./LoginLink";
+import LogoutLink from "./LogoutLink";
+export { LoginLink, LogoutLink };
